backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.4.0
setuptools>=65.0.0

# Testing
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
